Load the gap_wide dataset for tidyr lesson
Added R code to load the gap_wide dataset in Data Manipulation with tidyr.
This dataset is referenced in the lesson without being loaded.
Loading the data with read.csv() causes problems later (see /issues/94) because
columns are loaded as Factor not chr.

Loading the dataset with stringsAsFactors = FALSE (referenced in comments on line 26 of the Rmd)
fixes the problem and the all.equal() function works as intended.
Columns 37 and 38 are still imported as int instead of num, but this does not lead to problems.