Added na.rm in calculations of mean and sd
Since the mutate statement has set certain values in the new variable gdp_billion to NA, calculating the mean and sd of this variable will give NA for e.g. Africa in 1992, which is less helpful. WIth the na.rm, these calculated measures have values. Including the na.rm in the mean and sd of gdpPercap and pop is simply to drive the point home.